# docs/api-spec/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend service for eev",
        "title": "eev-backend",
        "contact": {
            "name": "Alen Alex",
            "url": "http://alenalex.me",
            "email": "contact@alenalex.me"
        },
        "version": "1.0"
    },
    "host": "e.alenalex/me",
    "basePath": "/api/",
    "paths": {
        "/auth/google/callback": {
            "post": {
                "description": "Validates Google OAuth code and state, creates or retrieves user, sets access_token and refresh_token HttpOnly cookies.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth callback",
                "parameters": [
                    {
                        "description": "Google OAuth callback request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User authenticated successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleCallbackResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or state",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "OAuth validation failed",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/google/login": {
            "get": {
                "description": "Generates OAuth state, stores it in HttpOnly cookie, and returns Google authorization URL.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Initiate Google OAuth login",
                "responses": {
                    "200": {
                        "description": "Returns Google OAuth URL",
                        "schema": {
                            "$ref": "#/definitions/handlers.GoogleLoginResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to generate state",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns the currently authenticated user based on the access_token HttpOnly cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get current logged-in user info",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user",
                        "schema": {
                            "$ref": "#/definitions/handlers.MeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me/email-history": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Returns the past email targets of the user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get the past email targets of the user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search email substring",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user history",
                        "schema": {
                            "$ref": "#/definitions/handlers.TargetUserEmailResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.GoogleCallbackRequest": {
            "type": "object",
            "required": [
                "code",
                "state"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "handlers.GoogleCallbackResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "handlers.GoogleLoginResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "handlers.MeResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.TargetUserEmailResponse": {
            "type": "object",
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.TargetUserEmails"
                    }
                }
            }
        },
        "internal.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "services.TargetUserEmails": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "is_starred": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "This API uses cookie-based authentication. Clients must send the `access_token` cookie for auth.",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        },
        "RefreshTokenAuth": {
            "description": "Refresh token cookie (HttpOnly). Sent only to refresh endpoint.",
            "type": "apiKey",
            "name": "refresh_token",
            "in": "cookie"
        }
    },
...
}